Added ParallelEvaluator instead of ParallelEvaluationStep
