+++ geneticengine/algorithms/gp/operators/crossover.py
@@ -6,7 +6,7 @@ from geneticengine.core.problems import Problem
 from geneticengine.core.random.sources import Source
 from geneticengine.core.representations.api import CrossoverOperator
 from geneticengine.core.representations.api import Representation
-from geneticengine.evaluators import Evaluator
+from geneticengine.core.evaluators import Evaluator
 
 
 class GenericCrossoverStep(GeneticStep):
