+++ geneticengine/algorithms/gp/operators/elitism.py
@@ -6,7 +6,7 @@ from geneticengine.algorithms.gp.structure import GeneticStep
 from geneticengine.core.problems import Problem
 from geneticengine.core.random.sources import Source
 from geneticengine.core.representations.api import Representation
-from geneticengine.evaluators import Evaluator
+from geneticengine.core.evaluators import Evaluator
 
 
 class ElitismStep(GeneticStep):
