+++ geneticengine/algorithms/gp/operators/mutation.py
@@ -9,7 +9,7 @@ from geneticengine.core.problems import Problem
 from geneticengine.core.random.sources import Source
 from geneticengine.core.representations.api import MutationOperator
 from geneticengine.core.representations.api import Representation
-from geneticengine.evaluators import Evaluator
+from geneticengine.core.evaluators import Evaluator
 
 
 class GenericMutationStep(GeneticStep):
