+++ geneticengine/algorithms/gp/operators/novelty.py
@@ -5,7 +5,7 @@ from geneticengine.algorithms.gp.structure import GeneticStep
 from geneticengine.core.problems import Problem
 from geneticengine.core.random.sources import Source
 from geneticengine.core.representations.api import Representation
-from geneticengine.evaluators import Evaluator
+from geneticengine.core.evaluators import Evaluator
 
 
 class NoveltyStep(GeneticStep):
