+++ geneticengine/algorithms/gp/structure.py
@@ -6,7 +6,7 @@ from geneticengine.algorithms.gp.individual import Individual
 from geneticengine.core.problems import Problem
 from geneticengine.core.random.sources import Source
 from geneticengine.core.representations.api import Representation
-from geneticengine.evaluators import Evaluator
+from geneticengine.core.evaluators import Evaluator
 
 
 class PopulationInitializer(abc.ABC):
