+++ geneticengine/algorithms/heuristics.py
@@ -8,7 +8,7 @@ from geneticengine.core.grammar import Grammar
 from geneticengine.core.problems import Problem
 from geneticengine.core.random.sources import RandomSource
 from geneticengine.core.representations.api import Representation
-from geneticengine.evaluators import Evaluator
+from geneticengine.core.evaluators import Evaluator
 
 
 class Heuristics(ABC):
