+++ geneticengine/core/representations/grammatical_evolution/dynamic_structured_ge.py
@@ -21,7 +21,7 @@ from geneticengine.core.utils import get_generic_parameter
 from geneticengine.core.utils import is_generic
 from geneticengine.core.utils import is_generic_list
 from geneticengine.core.utils import strip_annotations
-from geneticengine.evaluators import Evaluator
+from geneticengine.core.evaluators import Evaluator
 from geneticengine.exceptions import GeneticEngineError
 from geneticengine.metahandlers.base import is_metahandler
